src/rag-core/pyproject.toml
[project]
name = "rag-core"
version = "0.1.0"
description = "Core RAG service for document ingestion and generation."
authors = [{ name = "Kilyan"}]
requires-python = ">=3.12"
dependencies = [
    # Core RAG & LangChain
    "langchain-classic==1.0.0",
    "langchain-core==1.0.4",
    "langchain-community==0.4.1",
    "langchain-openai==1.0.2",
    "langchain-postgres==0.0.16",
    "fastembed==0.7.2",
    
    # Data Handling & Storage
    "boto3==1.40.75",
    "psycopg==3.2.12",
    "sqlalchemy==2.0.44",
    "unstructured[docx]==0.18.18",
    "pypdf==6.1.1",
    "markdown==3.10",
    "tiktoken==0.12.0",
    
    # API Framework
    "uvicorn==0.37.0",
    "fastapi==0.118.0",
    "python-multipart==0.0.20",
    
    # Models & Config
    "pydantic==2.11.10",
    "pydantic[email]==2.11.10",
    "pydantic-settings==2.11.0",

    # Security
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.3.0",
    "python-jose[cryptography]==3.5.0",
    "cryptography==46.0.3"
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "black==25.9.0",
    "ruff==0.14.1",
]

[tool.black]
line-length = 170

[tool.ruff]
line-length = 170

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
]